Refactor entries UI into modular layer
